[PR] ✨ feat: 투두, 루틴 리마인더 알림 발송 구현

+++ src/main/java/com/opu/opube/feature/notification/command/application/dto/response/NotificationResponse.java
@@ -32,7 +32,7 @@ public class NotificationResponse {
             example = "55",
             nullable = true
     )
-    private Integer linkedContentId;
+    private Long linkedContentId;
 
     @Schema(description = "읽음 여부", example = "false")
     private Boolean read;
